requirements.txt
asgiref==3.5.0
backports.zoneinfo==0.2.1
Brotli==1.0.9
click==8.0.4
colorama==0.4.4
dash==1.20.0
dash-core-components==1.16.0
dash-html-components==1.1.3
dash-renderer==1.9.1
dash-table==4.11.3
defusedxml==0.7.1
diff-match-patch==20200713
Django==4.0.2
django-database-view==0.3.0
django-environ==0.8.1
django-filter==21.1
django-import-export==2.7.1
django-plotly-dash==1.6.6
dpd-components==0.1.0
et-xmlfile==1.1.0
Flask==2.0.3
Flask-Compress==1.10.1
future==0.18.2
itsdangerous==2.1.0
Jinja2==3.0.3
MarkupPy==1.14
MarkupSafe==2.1.0
mysqlclient==2.1.0
numpy==1.22.2
odfpy==1.4.1
openpyxl==3.0.9
pandas==1.4.1
plotly==5.6.0
psycopg2==2.9.3
python-dateutil==2.8.2
pytz==2021.3
PyYAML==6.0
six==1.16.0
sqlparse==0.4.2
tablib==3.2.0
tenacity==8.0.1
tzdata==2021.5
Werkzeug==2.0.3
xlrd==2.0.1
xlwt==1.3.0
